pyproject.toml
[project]
name = "oneshot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "boto3>=1.42.7",
    "fastapi>=0.123.8",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
